requirements.txt
Django==4.1.3
mysqlclient==2.2.0
Pillow==9.5.0
django-dotenv==1.4.2
django-environ==0.10.0
django-jazzmin==2.6.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2==2.9.7
azure-core==1.29.3
azure-storage-blob==12.17.0
django-storages==1.13.2
